payloads: add networks list to READY
The Stats structure in payloads/stats.go includes a list of networks
present on the node.  This should also be reported in the ssntp.READY
status frame so that initial state is known on a node's first check-in,
which is where the ready command is used.

The information is available later since more detailed status, including
details on all running instances, is reported through the payloads.Stat
in ssntp ssntp.STATS command frames.  But that structure is deliberately
not interrogated by the scheduler because it is potentially tremendously
huge and is primarily targeted at the controller and higher layers of
the stack.  Enabling scheduler to know about a node's networks will enable
(re)scheduling of CNCI workloads to a suitable network node.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>