ciao-scheduler: use bool internally for net node flag
Since the payloads.Start resource type only has an integer value, zero and
one are used to encode the false and true of a request for a network node.
Internally it's easier if this is a bool.  This patch refactors scheduler
internally to use a bool and should amount to no functional change.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>