payloads: add optional string formatted value to resource request type
Initially we argued everything we'd want to request for a schedulable
resource would have a quantity associated with it.  Within that it's
easy enough to encode a boolean true/false with 1/0.  But we can't
encode something like a subnet in a way that's sensible.  Sure we could
convert a dotted-quad or a colon-hex to an integer, but that's going to
lead to a poorly readable workload definition yaml file for a human.

The first user of this field will be code to enable (re)scheduling of
CNCI's onto network nodes with connectivity to a needed subnet.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>