testutil: don't litter result channels in timeout tests
The channel timeout tests are leaving command/error/event/status results
on the test channels.  They need to each clean up after themselves,
otherwise the channel is pre-seeded with an event which can cause
another test to incorrectly proceed instead of waiting synchronously for
a result.

As the proverb says: Channels orchestrate; mutexes serialize.  We're
hoping for a certain amount of serialization and aren't getting it
because the orchestration is faulty when it leaves unhandled messages on
the channel.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>